+++ src/zombie.h
@@ -40,7 +40,7 @@ static Attribute attribute = {
   }
 };
 
-enum Race race = human;
+static enum Race race = human;
 
 static Window *s_main_window;
 
@@ -51,5 +51,22 @@ static TextLayer *s_location_layer;
 static GFont s_time_font;
 static GFont s_date_font;
 static GFont s_location_font;
+
+static char point_counter_buffer[32];
+static int point_counter = 0;
+
+// Function prototypes
+// main.c
+int main(void);
+// static void init(void);
+// static void deinit(void);
+
+// window.c
+void main_window_load(Window *window);
+void main_window_unload(Window *window);
+
+// handlers.c
+void select_click_handler(ClickRecognizerRef recognizer, void *context);
+void click_config_provider(void *context);
 
 #endif
